Fix module conflect
